consistent react function calling & organized interfaces
