--- client/src/pages/styles.ts
@@ -1,10 +1,0 @@
-import { styled, Box } from '@mui/system';
-
-export const TableBox = styled(Box)({
-  display: 'flex',
-  flexDirection: 'column',
-  justifyContent: 'center',
-  alignItems: 'center',
-  flex: 1,
-  paddingBottom: '5vh',
-});
